feature: Redis 테스트 코드 작성 - #2 디렉토리 구성 변
