+++ src/main/java/com/sentryq/thunder/domain/event/constant/ApplyResult.java
@@ -1,0 +1,8 @@
+package com.sentryq.thunder.domain.event.constant;
+
+public enum ApplyResult {
+    ;
+    public static final boolean SOLD_OUT = false;
+    public static final boolean DUPLICATE = false;
+    public static final boolean SUCCESS = true;
+}
