+++ src/main/java/com/sentryq/thunder/global/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.sentryq.thunder.config;
+package com.sentryq.thunder.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
