+++ src/main/java/com/sentryq/thunder/global/config/RedisKeyHelper.java
@@ -1,4 +1,4 @@
-package com.sentryq.thunder.api.utils;
+package com.sentryq.thunder.global.config;
 
 public class RedisKeyHelper {
